Issuance issue fix
