+++ backend/carbon-services-lib/src/data-importer/data-importer.module.ts
@@ -10,7 +10,7 @@ import { ProgrammeModule } from '../shared/programme/programme.module';
 import { TypeOrmConfigService } from '../shared/typeorm.config.service';
 import { Company } from '../shared/entities/company.entity';
 import { ProgrammeDocument } from '../shared/entities/programme.document';
-import { UtilModule } from '../shared/util/util.module';
+import { AnnualReportModule } from 'src/shared/annualreport/annual-report.module';
 
 @Module({
   imports: [
@@ -30,7 +30,7 @@ import { UtilModule } from '../shared/util/util.module';
     ProgrammeModule,
     CompanyModule,
     UserModule,
-    UtilModule
+    AnnualReportModule
     
   ],
   providers: [Logger, DataImporterService],
