+++ backend/carbon-services-lib/src/data-importer/data-importer.service.ts
@@ -11,7 +11,7 @@ import { CompanyService } from '../shared/company/company.service';
 import { Company } from '../shared/entities/company.entity';
 import { AnnualReportImport } from './importers/annual-report.service';
 import { ProgrammeDocument } from '../shared/entities/programme.document';
-import { AnnualReportGen } from '../shared/util/annual.report.gen';
+import { AnnualReportGen } from '../shared/annualreport/annual.report.gen';
 @Injectable()
 export class DataImporterService {
   constructor(
